Fix incorrect equals and hashCode for SearchResult
In this commit we change existing sealed SearchResult class and move
all its data to subclasses, which are now data classes, so equals and
hashCode are generated automatically. Additionally, we remove positions
from RemoveSearchResult class, because they are not needed there. Also
during testing we fix the situation, when file is already deleted, but
we try to update search results with file reading.